pyproject.toml
[project]
name = "etl_coro_example"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=36.1.1",
    "pqdm>=0.2.0",
    "psycopg>=3.2.5",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "types-tqdm>=4.67.0.20250301",
]

[tool.mypy]
strict = true
plugins = "pydantic.mypy"
